requirements.txt
torch>=2.0.0
transformers>=4.31.0
peft>=0.4.0
datasets>=2.10.0
pillow>=9.0.0
scikit-learn>=1.0.0
llama-cpp-python>=0.2.0
tqdm>=4.64.0
pandas>=1.3.0
matplotlib>=3.5.0